Separate dashboard tests; update DashboardTab and DashboardTabs implementations
